Extract the CLI commands from app.go into commands.go
